+++ examples/resources/scalr_module_namespace/resource.tf
@@ -1,0 +1,4 @@
+resource "scalr_module_namespace" "example" {
+  name      = "example-namespace"
+  is_shared = true
+}